+++ apps/service/src/pages/waitingDetail/WaitingDetailPage.tsx
@@ -1,5 +1,5 @@
-import { useEffect, useState } from "react";
-import { useNavigate, useParams, useLocation } from "react-router-dom";
+import { useEffect } from "react";
+import { useNavigate, useParams } from "react-router-dom";
 import * as S from "./WaitingDetailPage.styled";
 import BoothCardDetail from "@components/boothCard/boothCardDetail";
 import BottomButton from "@components/bottomButton/BottomButton";
@@ -12,15 +12,14 @@ import { useModal } from "@linenow/core/hooks";
 import WaitingDetailMap from "./_components/WaitingDetailMap";
 import { postWaitingCancel } from "@apis/domains/waiting/postWaitingCancel";
 import { modalCancelWaiting } from "@components/modal/waiting";
+import useToastFromLocation from "@hooks/useToastFromLocation";
 // import useAnimation from "./hooks/useAnimation";  // 주석 처리
 
 const WaitingDetailPage = () => {
   const navigate = useNavigate();
-  const location = useLocation();
   const params = useParams<{ waitingID: string }>();
   const waitingID = parseInt(params.waitingID || "0", 10);
-  const [showToast, setShowToast] = useState(false);
-  const [toastMessage, setToastMessage] = useState("");
+  const { showToast, toastMessage } = useToastFromLocation();
 
   const { data: waitingDetail, isLoading } = useGetWaiting(waitingID);
   const { data: waitingBooth } = useGetWaitingBooth(waitingID);
@@ -28,18 +27,6 @@ const WaitingDetailPage = () => {
 
   // 애니메이션 관련 코드 주석 처리
   // const { fadeInCard, slideUpCard, showRest, showToast } = useAnimation(location.state?.withAnimation);
-
-  useEffect(() => {
-    if (location.state?.showToast) {
-      setShowToast(true);
-      setToastMessage(location.state.toastMessage);
-
-      navigate(location.pathname, {
-        replace: true,
-        state: {},
-      });
-    }
-  }, [location.state]);
 
   const cancelWaiting = () => {
     if (waiting?.waitingID !== undefined) {
